backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.12.5
pydantic-settings==2.3.1
python-multipart==0.0.9
requests==2.32.3
pillow==12.1.0
pytesseract==0.3.10
pdf2image==1.17.0
pypdf==4.3.1
reportlab==4.0.7
sqlalchemy==2.0.30
alembic==1.13.2
psycopg2-binary==2.9.11

pandas>=2.0.0
openpyxl>=3.0.0  # For Excel file support

sendgrid>=6.10.0
twilio>=8.0.0
